+++ lib/pages/others/help-screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_i18n/flutter_i18n.dart';
 
-import 'package:moblie/utils/settings.dart';
 import 'package:moblie/widgets/ui/list_cell.dart';
 import 'package:moblie/widgets/ui/text/subtitle.dart';
 import 'package:moblie/widgets/ui/text/title.dart';
@@ -11,44 +11,48 @@ class HelpScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Solución de problemas'),
+        title: Text(FlutterI18n.translate(context, 'troubleshooting.title')),
         centerTitle: true,
       ),
       body: Scrollbar(
         child: ListView(
           children: <Widget>[
-            CommonTitle('No me aparecen palabras'),
-            SubTitle('Siempre tengo el mensaje: $onServerConnection'),
+
+            CommonTitle(FlutterI18n.translate(context, 'troubleshooting.no_words.title')),
+            SubTitle(FlutterI18n.translate(context, 'troubleshooting.no_words.subtitle')),
             ListCell(
               leading: Image.asset(
                 'assets/wifi.png',
                 height: 50.0,
                 width: 50.0,
               ),
-              title: ('Conexión a internet.'),
-              subtitle: ('Una razón podría ser que conexión a internet puede estar fallando o es muy lenta, por favor comprueba la conexión a internet. Si nofunciona, intenta más tarde.'),
+              title: FlutterI18n.translate(context, 'troubleshooting.no_words.response_one'),
+              subtitle: FlutterI18n.translate(context, 'troubleshooting.no_words.solution_one'),
               onTap: () {},
             ),
-            CommonTitle('¡No puedo escuchar nada!'),
-            SubTitle('En algunos dispositivos, la aplicación puede mostrar un comportamiento extraño, pero puedes solucionarlo con la siguiente guía.'),
+
+            CommonTitle(FlutterI18n.translate(context, 'troubleshooting.no_volume.title')),
+            SubTitle(FlutterI18n.translate(context, 'troubleshooting.no_volume.subtitle')),
+
             ListCell(
-              title: ('Se paciente.'),
-              subtitle: ('Esta aplicación está en fase beta y podría tardar un poco en responder, puede tardar hasta 20 segundos, por favor se paciente.'),
-              onTap: () {},
-            ),
-            ListCell(
-              title: ('Comprueba el volumen.'),
-              subtitle: ('Podría ser obvio, pero ¿has intentado aumentar el volumen?'),
+              title: FlutterI18n.translate(context, 'troubleshooting.no_volume.response_one'),
+              subtitle: FlutterI18n.translate(context, 'troubleshooting.no_volume.solution_one'),
               onTap: () {},
             ),
             ListCell(
-              title: ('Modo no molestar.'),
-              subtitle: ('Es posible que tenga activado el modo "No molestar". Intente desactivarlo.'),
+              title: FlutterI18n.translate(context, 'troubleshooting.no_volume.response_two'),
+              subtitle: FlutterI18n.translate(context, 'troubleshooting.no_volume.solution_two'),
+              onTap: () {},
+            ),
+
+            ListCell(
+              title: FlutterI18n.translate(context, 'troubleshooting.no_volume.response_three'),
+              subtitle: FlutterI18n.translate(context, 'troubleshooting.no_volume.solution_three'),
               onTap: () {},
             ),
             ListCell(
-              title: ('Última revisión.'),
-              subtitle: ('Si nada de esto funciona, tienes dos opciones, puede que tu teléfono tenga un altavoz que no funciona o tu dispositivo no escompatible con la aplicación.'),
+              title: FlutterI18n.translate(context, 'troubleshooting.no_volume.response_four'),
+              subtitle: FlutterI18n.translate(context, 'troubleshooting.no_volume.solution_four'),
               onTap: () {},
             ),
           ]