+++ lib/pages/palabra-guardada-screen.dart
@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_i18n/flutter_i18n.dart';
+
+import 'package:scoped_model/scoped_model.dart';
 import 'package:moblie/model/main.dart';
+
 import 'package:moblie/pages/screens/error-screen.dart';
 import 'package:moblie/pages/search-screen.dart';
-import 'package:moblie/utils/settings.dart';
+
 import 'package:moblie/widgets/palabra-guardada/palabra-guardada.dart';
-import 'package:scoped_model/scoped_model.dart';
 
 class PalabraGuardadaScreen extends StatefulWidget {
   final MainModel model;
@@ -28,9 +31,9 @@ class _PalabraGuardadaScreenState extends State<PalabraGuardadaScreen> {
       builder: (BuildContext context, Widget child, MainModel model) {
 
         Widget content = ErrorScreen(
-          message: onNoSavedWords,
+          message: FlutterI18n.translate(context, 'error_message.no_saved_words.message'),
           pathImage: 'assets/box.png',
-          fixMessage: onNoSavedWordsFix,
+          fixMessage: FlutterI18n.translate(context, 'error_message.no_saved_words.solution')
         );
 
         if (model.allPalabrasGuardadas.length > 0 && !model.palabrasGuardadasIsLoading) {
@@ -48,19 +51,22 @@ class _PalabraGuardadaScreenState extends State<PalabraGuardadaScreen> {
   Widget build(BuildContext context) {
     return Scaffold (
       appBar: AppBar(
-        title: Text(savedSection),
+        title: Text(FlutterI18n.translate(context, 'saved_words.title')),
         centerTitle: true,
         actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.search),
-            onPressed: () => Navigator
-              .of(context)
-              .push(searchPalabras(
-                context, 
-                widget.model.allPalabrasGuardadas, 
-                widget.model
-              )),
-          )
+          Tooltip(
+            message: FlutterI18n.translate(context, 'saved_words.appbar.icons.search_tooltip'),child: IconButton(
+              icon: Icon(Icons.search),
+              onPressed: () => Navigator
+                .of(context)
+                .push(searchPalabras(
+                  context, 
+                  widget.model.allPalabrasGuardadas, 
+                  widget.model
+                )),
+            ),
+          ),
+
         ],
       ),
       body:  _buildMainContent(),